Ai-Service/requirements.txt
fastapi
uvicorn[standard]
google-genai
google-generativeai
python-dotenv